# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: A1 Agent Engine Platform API
  version: 1.0.0
  description: >
    Comprehensive API specification for the A1 Agent Engine — an enterprise agentic PaaS with
    Temporal-based durable workflows, multi-tenancy, and MCP integration.

    The platform operates in two distinct MCP modes:
    - **Platform as MCP Client** (mcp-registry) — manages connections to 3rd-party MCP servers
    - **Platform as MCP Server** (mcp-server) — exposes platform capabilities via MCP protocol
  contact:
    name: Platform Team
    email: support@agentplatform.local
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development API Gateway
  - url: http://localhost:8081
    description: Local development Workflow Initiator
  - url: http://localhost:8089
    description: Local development Admin API
  - url: http://localhost:8090
    description: Local development MCP Registry (Tenant MCP Servers)
  - url: http://localhost:8091
    description: Local development MCP Server (Platform Capabilities)
...
components:
  securitySchemes:
    HmacAuth:
      type: apiKey
      name: X-Signature
      in: header
      description: >
        HMAC-SHA256 signature. Format: `X-Signature: sha256=<hex>`.
        Paired with `X-Timestamp: <unix-epoch>` header. Window tolerance: 300s.
    AdminKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: custom
      description: >
        Admin API key authentication. Either `Authorization: Bearer <key>` or `X-Admin-Key: <key>` header.
        Token must match `ADMIN_API_KEY` environment variable.
    McpTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: sha256
      description: >
        MCP Server authentication. SHA-256 hashed bearer token from `mcp_tokens` table.
        `Authorization: Bearer <token>` header. Tokens are looked up and validated against tenant.
  schemas:
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code (e.g., "INVALID_REQUEST", "UNAUTHORIZED", "RESOURCE_NOT_FOUND")
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Optional structured error details
          additionalProperties: true
...
paths:
...
  /mcp:
    post:
      tags:
        - MCP Server
      summary: MCP JSON-RPC 2.0 endpoint
      description: >
        The MCP protocol endpoint. External MCP clients (Claude Desktop, etc.)
        connect here using JSON-RPC 2.0 over HTTP.

        Supported methods:
        - `initialize` — Start MCP session
        - `tools/list` — Discover available tools (skills + platform APIs + KG tools)
        - `tools/call` — Invoke a tool
      x-audience: external
      x-websocket: false
      security:
        - McpTokenAuth: []
      operationId: handleMCP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - jsonrpc
                - method
                - id
              properties:
                jsonrpc:
                  type: string
                  enum: ['2.0']
                method:
                  type: string
                  enum: [initialize, tools/list, tools/call]
                params:
                  type: object
                  description: Method-specific parameters
                  additionalProperties: true
                id:
                  type: integer
                  description: Request ID (must be echoed in response)
      responses:
        '200':
          description: MCP JSON-RPC response
          content:
            application/json:
              schema:
                type: object
                required:
                  - jsonrpc
                  - id
                properties:
                  jsonrpc:
                    type: string
                    enum: ['2.0']
                  result:
                    type: object
                    description: Success result object (method-dependent)
                    additionalProperties: true
                  error:
                    type: object
                    description: JSON-RPC error object (if applicable)
                    properties:
                      code:
                        type: integer
                      message:
                        type: string
                      data:
                        type: object
                  id:
                    type: integer
        '400':
          description: Malformed JSON-RPC request
        '401':
          description: Unauthorized (invalid or missing token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...